+++ Cargo.toml
@@ -24,10 +24,12 @@ axum = { version = "0.8.0-alpha.1" }
 bincode = { version = "1.3.3" }
 clap = { version = "4.5.17" }
 config = { version = "0.14.0" }
+deadpool-diesel = { version = "0.6.1" }
+diesel = { version = "2.2.6" }
+diesel_migrations = { version = "2.2.0" }
 glob = { version = "0.3.1" }
 graphile_worker = { version = "0.8.0" }
 itertools = { version = "0.13.0" }
-lazy_static = { version = "1.5.0" }
 log = { version = "0.4.22" }
 opentelemetry = { version = "0.27.0" }
 opentelemetry-otlp = { version = "0.27.0" }
@@ -46,8 +48,10 @@ test-log = { version = "0.2.16" }
 testcontainers-modules = { version = "0.11.3" }
 thiserror = { version = "2.0.3" }
 tokio = { version = "1.39.3" }
+tower = { version = "0.5.1" }
 tower-http = { version = "0.6.1" }
 tracing = { version = "0.1.40" }
 tracing-bunyan-formatter = { version = "0.3.9" }
 tracing-opentelemetry = { version = "0.28.0" }
 tracing-subscriber = { version = "0.3.18" }
+uuid = { version = "1.11.0" }
