+++ crates/adapter/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod http;
 pub mod in_memory;
+pub mod postgres;
