+++ crates/adapter/src/postgres/config.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug, Clone)]
+pub struct DBConfig {
+    pub url: String,
+    pub max_size: usize,
+}
