+++ crates/adapter/src/postgres/diesel.toml
@@ -1,0 +1,9 @@
+# For documentation on how to configure this file,
+# see https://diesel.rs/guides/configuring-diesel-cli
+
+[print_schema]
+custom_type_derives = ["diesel::query_builder::QueryId"]
+file = "schema.rs"
+
+[migrations_directory]
+dir = "migrations"
