+++ crates/adapter/src/postgres/mod.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod models;
+pub mod schema;
+pub mod session_db;
