+++ crates/adapter/src/postgres/models/mod.rs
@@ -1,0 +1,1 @@
+pub mod session;
