+++ crates/core/Cargo.toml
@@ -9,3 +9,4 @@ async-trait = { workspace = true }
 phantom = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 thiserror = { workspace = true }
+uuid = { workspace = true, features = ["v4", "serde"] }
