+++ crates/server/src/services/session.rs
@@ -1,27 +1,26 @@
-use std::string::ToString;
 use std::sync::Arc;
 
 use frog_core::entities::client::{ClientEntity, ClientId};
 use frog_core::entities::session::{SessionEntity, SessionId, SessionStatus};
 use frog_core::ports::session::SessionPort;
-use lazy_static::lazy_static;
-use phantom::fhe_function;
-use phantom::phantom_zone::{Crs, Ops, Param};
+use phantom::crs::Crs;
+use phantom::ops::Ops;
+use phantom::param::Param;
+use phantom::utils::fhe_function;
 use phantom_zone_evaluator::boolean::fhew::prelude::FheU64;
 use phantom_zone_evaluator::boolean::FheBool;
+use tokio::sync::Mutex;
 
 use crate::errors::AppError;
 use crate::errors::AppError::SessionError;
-
-lazy_static! {
-    static ref SESSION_ID: SessionId = SessionId("0".to_string());
-}
 
 pub struct SessionService {
     session: Arc<dyn SessionPort + Sync + Send>,
     phantom_param: Param,
     crs: Crs,
     participant_number: usize,
+
+    mutex: Arc<Mutex<bool>>,
 }
 
 impl SessionService {
@@ -36,16 +35,24 @@ impl SessionService {
             phantom_param,
             crs,
             participant_number,
+            mutex: Default::default(),
         }
     }
 
     pub async fn create(&self) -> Result<(), AppError> {
         self.session
             .create(SessionEntity::new(
-                SESSION_ID.clone(),
+                SessionId::try_from("f8e774bd-2f9d-4502-92ca-ac8b9c25868e")?,
                 self.phantom_param,
                 self.crs,
             ))
+            .await?;
+        Ok(())
+    }
+
+    pub async fn delete(&self) -> Result<(), AppError> {
+        self.session
+            .delete(SessionId::try_from("f8e774bd-2f9d-4502-92ca-ac8b9c25868e")?)
             .await?;
         Ok(())
     }
@@ -107,6 +114,7 @@ impl SessionService {
         client_id: ClientId,
         bs_key: Vec<u8>,
     ) -> Result<(), AppError> {
+        let guard = self.mutex.lock().await;
         let mut session_entity = self.session.get(session_id.clone()).await?;
         let client_entity = session_entity.client_info.get_mut(&client_id).unwrap();
         client_entity.bs_key_share = bs_key;
@@ -118,15 +126,16 @@ impl SessionService {
             .collect::<Vec<_>>();
         if bs_key_shares.len() == self.participant_number {
             let phantom_server = session_entity.phantom_server.as_mut().unwrap();
-            phantom_server.aggregate_bs_key_shares(
-                &bs_key_shares
-                    .iter()
-                    .map(|bytes| phantom_server.deserialize_bs_key_share(bytes).unwrap())
-                    .collect::<Vec<_>>(),
-            );
+            let mut bs_key_shares = bs_key_shares
+                .iter()
+                .map(|bytes| phantom_server.deserialize_bs_key_share(bytes).unwrap())
+                .collect::<Vec<_>>();
+            bs_key_shares.sort_by_key(|bs_key| bs_key.share_idx());
+            phantom_server.aggregate_bs_key_shares(&bs_key_shares);
             session_entity.status = SessionStatus::WaitingForArgument;
         }
         self.session.update(session_id, session_entity).await?;
+        let _ = guard;
         Ok(())
     }
 
@@ -136,6 +145,7 @@ impl SessionService {
         client_id: ClientId,
         data: Vec<u8>,
     ) -> Result<(), AppError> {
+        let guard = self.mutex.lock().await;
         let mut session_entity = self.session.get(session_id.clone()).await?;
         let client_entity = session_entity.client_info.get_mut(&client_id).unwrap();
         client_entity.encrypted_data = data;
@@ -168,6 +178,7 @@ impl SessionService {
             session_entity.status = SessionStatus::Done;
         }
         self.session.update(session_id, session_entity).await?;
+        let _ = guard;
         Ok(())
     }
 }
